Refactor ContactsRedux and ContactsSaga into feature archtecture
