+++ App/Redux/RootReducer.ts
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux'
 import { reducer as accountReducer } from './AccountRedux'
 import { reducer as authReducer } from './AuthRedux'
 import { reducer as cameraRollReducer } from './CameraRollRedux'
-import { reducer as contactsReducer } from './ContactsRedux'
+import { contactsReducer } from '../features/contacts'
 import { reducer as photoViewingReducer } from './PhotoViewingRedux'
 import { reducer as prefrencesReducer } from './PreferencesRedux'
 import { reducer as notificationsReducer } from './NotificationsRedux'
