+++ App/Redux/Types.ts
@@ -5,7 +5,7 @@ import RootReducer from './RootReducer'
 import { AccountAction } from './AccountRedux'
 import { AuthAction } from './AuthRedux'
 import { CameraRollAction } from './CameraRollRedux'
-import { ContactsAction } from './ContactsRedux'
+import { ContactsAction } from '../features/contacts'
 import { PhotoViewingAction } from './PhotoViewingRedux'
 import { PreferencesAction } from './PreferencesRedux'
 import { NotificationsAction } from './NotificationsRedux'
