+++ App/features/contacts/index.ts
@@ -1,0 +1,6 @@
+import * as contactsActions from './actions'
+import contactsReducer, { ContactsState, ContactsAction } from './reducer'
+import * as contactsSelectors from './selectors'
+import contactsSaga from './sagas'
+
+export { contactsActions, contactsReducer, contactsSelectors, ContactsState, ContactsAction, contactsSaga }
